remove deprecated methods
