+++ src/main/scala/BintrayRepo.scala
@@ -55,9 +55,9 @@ case class BintrayRepo(credential: BintrayCredentials, org: Option[String], repo
       // warn the user that bintray expects maven published artifacts to be published to the `maven` repo
       // but they have explicitly opted into a publish style and/or repo that
       // deviates from that expecation
-      if ("maven" == repo && !mvnStyle) println(
+      if (Bintray.defaultMavenRepository == repo && !mvnStyle) println(
         "you have opted to publish to a repository named 'maven' but publishMavenStyle is assigned to false. This may result in unexpected behavior")
-      Opts.resolver.publishTo(bty, pkg, vers, mvnStyle, isSbtPlugin, isRelease)
+      Bintray.publishTo(bty, pkg, vers, mvnStyle, isSbtPlugin, isRelease)
     }
 
   def release(packageName: String, vers: String, log: Logger): Unit =
